requirements.txt

aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
branca==0.7.0
certifi==2023.7.22
channels==3.0.5
charset-normalizer==3.3.2
Django==4.2.6
django-filter==23.3
folium==0.15.0
frozenlist==1.4.0
geographiclib==2.0
geoip2==4.7.0
geopy==2.4.0
greenlet==3.0.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
maxminddb==2.5.1
multidict==6.0.4
numpy==1.26.1
Pillow==10.1.0
requests==2.31.0
SQLAlchemy==1.4.22
sqlparse==0.4.4
urllib3==2.0.7
yarl==1.9.2
GDAL==3.8.1
